+++ ReactWindows/ReactNative/Shell/MainReactPackage.cs
@@ -1,7 +1,7 @@
 using ReactNative.Animated;
 using ReactNative.Bridge;
 using ReactNative.Modules.AppState;
-using ReactNative.Modules.CameraRoll;
+using ReactNative.Modules.Camera;
 using ReactNative.Modules.Clipboard;
 using ReactNative.Modules.Core;
 using ReactNative.Modules.Dialog;
@@ -49,7 +49,7 @@ namespace ReactNative.Shell
             {
                 new AppStateModule(reactContext),
                 new AsyncStorageModule(),
-                new CameraRollModule(reactContext),
+                new CameraRollManager(reactContext),
                 new ClipboardModule(),
                 new DialogModule(reactContext),
                 new ImageLoaderModule(),
